[Profiler] Make KinetoEvent a view of Result (Part 2, python and stacks)
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81320

The semantics of python in KinetoEvent are sort of crazy because values can come either from metadata captured by JIT or from the python tracer. Things aren't helped by the fact that we encode the python hierarchy in the chrome trace for tensorboard, so we have to do a tree traversal with skips based on types.

By simply constructing the Python stack in KinetoEvent's ctor we're able to skip a lot of the complexity in `EventFieldsVisitor` and just lean on `KinetoEvent.stack()`.
ghstack-source-id: 163104855

Differential Revision: [D37481561](https://our.internmc.facebook.com/intern/diff/D37481561/)